requirements.txt
python-dotenv==1.0.1
openai==1.30.1
langchain[all]==0.1.20
langchain_openai==0.1.7
langchain-experimental==0.0.58
sentence-transformers==2.7.0
tokenizers==0.19.1
unstructured[all-docs]==0.13.7
tiktoken==0.7.0
PyPDF2==3.0.1
chromadb==0.5.0
